add github workflow
This is based on the workflow from HTTP::Message. We allow Perl 5.8, so this is turned on. However, several of our xt/ tests require Perl 5.12 or 5.14.

We have to be creative in how we exclude 5.8 from running xt/ tests,
because 5.8 is larger than 5.12.

Also bump Test::More requirement from 0.94 (implicit) to 0.96 to allow
our subtests.