parse select inputs that have no name #2
In order to do this, I had to allow find_input to find inputs that have
no name. Before, we treated an undef in the name (actually, the
'selector') as an instruction not to match the name. To put option
elements into the correct surrounding select, we have to match them by
their name, including when there is no name.

Two of the tests I have added don't produce what I would expect. I think
this is another bug. It seems we include unnamed fields that have a
value in the query string, as long as they have a name attribute. This
leads to query strings such as 'name=value&=value'. The W3 specification
explicitly says that only element that have names should be submitted,
so I believe this behaviour is wrong.