normalize newlines in forms to CRLF
HTTP::Message < 7.00 normalized newlines in requests to CRLF. This has
been removed. Based on the [HTML
spec](https://html.spec.whatwg.org/#converting-an-entry-list-to-a-list-of-name-value-pairs)
and [URL spec](https://url.spec.whatwg.org/#urlencoded-serializing), the CRLF
normalization should be in the HTML layer. So the change in
HTTP::Message is correct, and HTML::Form should be handling the
normalization.

Add normalization of lone \r or \n into \r\n, as defined in the HTML
spec.

Older HTTP::Message versions will do their own normalization. That
will end up being a no-op, so we will still be compatible with those
older versions.